change Area interface from result based to value based

+++ src/main/java/unit/scalar/adjustment/Identity.java
@@ -28,7 +28,7 @@ package unit.scalar.adjustment;
  */
 public class Identity implements Adjustment {
     @Override
-    public final float adjusted(final float current) {
+    public final double adjusted(final double current) {
         return current;
     }
 }
